requirements.txt
dash==2.9.1
dash_auth==2.0.0
dash_bootstrap_components==1.4.1
dash_leaflet==0.1.23
pandas==1.5.3
psycopg2==2.9.5
python-dotenv==1.0.0
gunicorn
dash-tools
